Fix: rename login Btn and redirect use Effect
