+++ src/pages/login/redirect.tsx
@@ -3,8 +3,11 @@ import {useRouter} from 'next/router';
 export default function Redirect() {
   const router = useRouter();
   console.log(router);
-  const auth_code = router.query.code || '';
-  console.log(auth_code);
+  useEffect(() => {
+    const auth_code = router.query.code;
+    console.log(auth_code);
+  }, []);
+
   // useEffect()
   return <div></div>;
 }
